% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeirf.R
\name{computeirf}
\alias{computeirf}
\title{Compute impulse response functions for a VAR(P)}
\usage{
computeirf(Psi, Phi, H, cum)
}
\arguments{
\item{Psi}{Impact matrix, dimension \verb{N x R}.}

\item{Phi}{Array of VAR coefficient matrices, dimension \verb{N x N x P}.}

\item{H}{Horizon (number of periods) for which to compute IRFs.}

\item{cum}{Logical scalar or logical vector of length \code{N}. If \code{TRUE} for
variable \code{i}, the IRF is cumulated via \code{cumsum}. A single value is applied to all variables.}
}
\value{
An array:
\describe{
\item{irf}{Array of impulse responses, dimension \verb{H x N x R}. Row names
are labelled \code{0} to \code{H-1}.}
}
}
\description{
Recursively computes IRFs given an impact matrix and VAR coefficient array.
Optionally cumulates responses for selected variables.
}
\examples{
Psi <- matrix(c(1, 0.5), nrow = 2)
Phi <- array(0, dim = c(2, 2, 1))
Phi[, , 1] <- matrix(c(0.4, 0.1, 0, 0.3), 2, 2)
computeirf(Psi = Psi, Phi = Phi, H = 4, cum = FALSE)

}
