% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetiv-package.R
\docType{package}
\name{hetiv-package}
\alias{hetiv-package}
\title{hetiv: Heteroskedasticity and Proxy-Based IV Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for identifying structural shocks using heteroskedasticity- and proxy-based instrumental variable (IV) methods in event-study settings. Supports local projection impulse response estimation of Jorda (2005) \doi{10.1257/0002828053828518}, generalised weak instrument testing of Lewis and Mertens (2025) \doi{10.1093/restud/rdaf103}, recursive heteroskedasticity-IV identification of Burri and Kaufmann (2026b) \doi{10.1016/j.econlet.2026.113091}, and Kalman-filter shock extraction of Burri and Kaufmann (2026a).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dankaufmann.github.io/hetiv/}
  \item \url{https://github.com/dankaufmann/hetiv}
  \item Report bugs at \url{https://github.com/dankaufmann/hetiv/issues}
}

}
\author{
\strong{Maintainer}: Daniel Kaufmann \email{info@dankaufmann.com} [copyright holder]

Authors:
\itemize{
  \item Daniel Kaufmann \email{info@dankaufmann.com} [copyright holder]
  \item Marc Burri
  \item Valentin Grob
}

}
\keyword{internal}
