% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetiv-package.R
\docType{package}
\name{hetiv-package}
\alias{hetiv-package}
\title{hetiv: Heteroskedasticity and Proxy-Based IV Methods}
\description{
Tools for identifying structural shocks using heteroskedasticity- and proxy-based instrumental variable (IV) methods in event-study settings. Supports local projection impulse response estimation (Jordà, 2005), generalised weak instrument testing (Lewis and Mertens, 2025), and Kalman-filter shock extraction (Burri and Kaufmann, 2026).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dankaufmann.github.io/hetiv/}
  \item \url{https://github.com/dankaufmann/hetiv}
  \item Report bugs at \url{https://github.com/dankaufmann/hetiv/issues}
}

}
\author{
\strong{Maintainer}: Daniel Kaufmann \email{info@dankaufmann.com}

Authors:
\itemize{
  \item Marc Burri
  \item Valentin Grob
}

}
\keyword{internal}
